infosedd-synthetic/requirements.txt
setuptools==78.1.1
wheel==0.45.1
pip==25.1
asttokens==3.0.0
comm==0.2.3
debugpy==1.8.11
decorator==5.2.1
entrypoints==0.4
executing==2.2.0
nest_asyncio==1.6.0
packaging==25.0
parso==0.8.4
pickleshare==0.7.5
platformdirs==4.3.8
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyzmq==26.2.0
six==1.17.0
traitlets==5.14.3
typing_extensions==4.14.1
wcwidth==0.2.13
exceptiongroup==1.3.0
jedi==0.19.2
jupyter_core==5.8.1
matplotlib-inline==0.1.7
pexpect==4.9.0
prompt_toolkit==3.0.51
psutil==5.9.1
python-dateutil==2.9.0.post0
stack_data==0.6.3
tornado==6.1
ipython==8.37.0
jupyter-client==7.3.4
ipykernel==6.29.5
numpy==2.2.6
scipy==1.15.3
nvidia-cusparselt-cu12==0.6.3
mpmath==1.3.0
triton==3.3.1
sympy==1.14.0
nvidia-nvtx-cu12==12.6.77
nvidia-nvjitlink-cu12==12.6.85
nvidia-nccl-cu12==2.26.2
nvidia-curand-cu12==10.3.7.77
nvidia-cufile-cu12==1.11.1.6
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cublas-cu12==12.6.4.1
networkx==3.4.2
MarkupSafe==3.0.2
fsspec==2025.7.0
filelock==3.18.0
nvidia-cusparse-cu12==12.5.4.2
nvidia-cufft-cu12==11.3.0.4
nvidia-cudnn-cu12==9.5.1.17
Jinja2==3.1.6
nvidia-cusolver-cu12==11.7.1.2
torch==2.7.1
pyparsing==3.2.3
pillow==11.3.0
kiwisolver==1.4.8
fonttools==4.59.0
cycler==0.12.1
contourpy==1.3.2
matplotlib==3.10.3
pytz==2025.2
tzdata==2025.2
pandas==2.3.1
tqdm==4.67.1
PyYAML==6.0.2
propcache==0.3.2
multidict==6.6.3
lightning-utilities==0.15.0
idna==3.10
frozenlist==1.7.0
attrs==25.3.0
async-timeout==5.0.1
aiohappyeyeballs==2.6.1
yarl==1.20.1
aiosignal==1.4.0
aiohttp==3.12.14
torchmetrics==1.8.0
pytorch-lightning==2.5.2
antlr4-python3-runtime==4.9.3
omegaconf==2.3.0
positional-encodings==6.0.3
hydra-core==1.3.2
nvidia-ml-py==12.575.51
nvitop==1.5.2
urllib3==2.5.0
typing-inspection==0.4.1
smmap==5.0.2
pydantic_core==2.33.2
protobuf==6.31.1
click==8.2.1
charset-normalizer==3.4.2
certifi==2025.7.14
annotated-types==0.7.0
sentry-sdk==2.34.0
requests==2.32.4
pydantic==2.11.7
gitdb==4.0.12
GitPython==3.1.45
wandb==0.21.0